requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
markdown==3.5.1
python-dateutil==2.8.2
sqlalchemy
psycopg2-binary
python-dotenv
